Log agregado y los Long id desde service interface para pelicula, falta para personaje y genero

+++ src/main/java/com/besysoft/practica/services/interfaces/GeneroService.java
@@ -13,7 +13,7 @@ public interface GeneroService {
 
     Genero crearGenero(Genero genero) throws Exception;
 
-    Genero actualizarGenero(Genero genero, int id)throws Exception;
+    Genero actualizarGenero(Genero genero, Long id)throws Exception;
 
     void borrarGenero(Long id);
